// src/models/finecoRecord.ts
export class FinecoRecord {
    operazione: string;
    dataValuta: string;
    descrizione: string;
    titolo: string;
    isin: string;
    segno: string;
    quantita: number;
    divisa: string;
    prezzo: number;
    cambio: number;
    controvalore: number;
    commissioniFondiSwIngrUscita: number;
    commissioniFondiBancaCorrispondente: number;
    speseFondiSgr: number;
    commissioniAmministrato: number;
}